mlops/requirements.txt
# Implementation guide: see docs/LEVEL-2-MLOPS.md — "mlops/requirements.txt"
# Pinned versions from plan doc.

# Orchestration
apache-airflow==2.8.1
# Experiment tracking
mlflow==2.10.0
# Serving
fastapi==0.109.0
uvicorn==0.27.0
onnxruntime-gpu==1.17.0
# Geospatial
rasterio==1.3.9
geopandas==0.14.2
shapely==2.0.2
# ML frameworks
torch==2.2.0
scikit-learn==1.4.0
xgboost==2.0.3
# Data
numpy==1.26.3
pandas==2.2.0
# Storage
minio==7.2.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
redis==5.0.1
# Monitoring
prometheus-client==0.19.0
